+++ CHANGES.rst
@@ -17,6 +17,12 @@ Unreleased
 * **[FEATURE]** Added support for thread locking. See
   :meth:`~praw.objects.Submission.lock` and
   :meth:`~praw.objects.Submission.unlock`.
+* **[FEATURE]** Added support for comment stickying. Try
+  :meth:`~praw.objects.Comment.distinguish` with new parameter `sticky`.
+* **[FEATURE]** Methods :meth:`~praw.__init__.ReportMixin.hide` and
+   :meth:`~praw.__init__.UnauthenticatedReddit.get_info` now support long lists
+   of fullnames. They will divide the list up into small, requestable chunks
+   automatically.
 
 3.4.0 (2016-02-21)
 ------------------